backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
sqlalchemy==2.0.48
pandas==3.0.2
numpy==2.4.4
requests==2.33.1
python-multipart==0.0.22
email-validator==2.3.0
PyJWT==2.10.1
bcrypt==4.3.0
python-jose==3.5.0
firebase-admin==7.3.0
cachetools==7.0.5
influxdb-client==1.50.0
